rspq: adapt mixer to changes

+++ src/audio/mixer.c
@@ -129,7 +129,7 @@ struct {
 /** @brief Count of ticks spent in mixer RSP, used for debugging purposes. */
 int64_t __mixer_profile_rsp = 0;
 
-static uint8_t __mixer_overlay_id;
+static uint32_t __mixer_overlay_id;
 
 static inline int mixer_initialized(void) { return Mixer.num_channels != 0; }
